service-auth/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout>=4.0.2
cachelib==0.6.0
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cspycore==1.0.0
dnspython==2.2.1
Faker==13.7.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Session==0.4.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.21.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.1.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.36
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.2.2